Move `EtcdClient` to a separate file
